Updated solar hybridization example for EAGO v0.7.1
Note that in the lower_problem! and upper_problem! definitions, if the sense of the @NLobjective is changed to "Min", the "flcs" term in @NLobjective and the two opt._lower_objective_value/opt._upper_objective_value definitions also need to be negated. Base EAGO is assuming this is a minimization problem, so if the lower/upper problems return values consistent with a maximization problem, it will cause errors.